observer_bundle/requirements.txt
pandas==2.2.2
numpy==1.26.4
requests==2.32.3
psycopg2-binary==2.9.10
fastapi==0.115.0
sse-starlette==2.1.3
uvicorn==0.30.6
python-telegram-bot==21.6
python-dotenv==1.0.1
ccxt==4.4.1
pytest==8.3.3
pytest-mock==3.14.0
